+++ internal/fake/client.go
@@ -19,8 +19,8 @@ package fake
 import (
 	"time"
 
-	"github.com/jthomperoo/k8shorizmetrics/metrics/podmetrics"
-	autoscalingv2 "k8s.io/api/autoscaling/v2beta2"
+	"github.com/jthomperoo/k8shorizmetrics/v2/metrics/podmetrics"
+	autoscalingv2 "k8s.io/api/autoscaling/v2"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 )